Returned rig polling 'fm' to 'fmv'
'v' is generating errors with Icom rigs, but Kenwoods can not survive without it. If 'v' is missing And Kenwood user pushes A/B vfo button from rig from panel cqrlog continues to show previous vfo frequency.
This is a rigctrl "feature" of Kenwood that it needs user action to poll current vfo before asking frequency.

uRigControl.pas is not suitable for fixing this easy. So we just leave errors for Icom as they do no bigger harm.

Added uRigControl.pas and uRotControl.pas to project. I can not see no harm with this. Finding them is now easier from unit list.