Setting DX spot line info text to clipboard
When DX spot is selected with double (left) click and DX callsign is set to NewQSO same time the info part of DX spot line is applied to clipboard.
User can then set cursor to a NewQSO field, like "comment to QSO", and press Ctrl-V to store the info (IOTA, SOTA, WWFF etc.) in easy way.

Squashed commit of the following:

commit eea816e6e6d62f70cca3f1ca437e348fe2da5d86
Author: OH1KH <oh1kh@sral.fi>
Date:   Sun Mar 22 13:18:03 2020 +0200

    Changed info copy to double(left)click and updated help file

commit d462a1c086745783af1b65e069430f0dbb80e9e3
Author: OH1KH <oh1kh@sral.fi>
Date:   Sat Mar 21 13:16:59 2020 +0200

    add item in DXCluster right click menu top: 'select, copy info, add to newqso'