Fix 1xESC keys
Several forms fixed so that if active 1xESC will return focus to NewQSO/edtCall.
This has been default action as most of forms already act this way.

In addition return to NewQSO/edtCall is fixed so that if Contest form is active return to Contest/edtCall is made instead.

Form fRemind was not part of project. Added to project.