Fix to modes bandwith
Rigctld has mode setting option "0" that means use default bandwidth.( M USB 0 )
That is mentioned in man pages.
What is not in man pages (until now) is that option "-1" is accepted too without errors. It means "No change". WSJT-X propgram uses that, so it can be there for all rigs)

Not all rigs support this, but as some do it must be possible to enter also "-1" to preferences/modes/bandwiths
Fixed MinValue of bandwidth edits to -1