Bug fix for UTF8
This fixes bug (long lasting) with UTF-8 that appears with Laz2.0.10/fpc3.2.0 but has been lying there already before.

This makes old qsos special characters "garbage" when they are saved with program compiled with fpc3.0.4 or earler.

That is why I do NOT expect this pull request is accepted !!
This is just to show how things must be, and for testing.

This problem needs further thinking how this is commited with less trouble for users.