RIT and XIT modifications
"New QSO/Clear rit after save qso" with Icom IC7300 this clears also XIT to zero. It is good thing specially if XIT has been used for split up/down for qso.
How ever all rig brands and models may not cear XIT when RIT is cleared. Therefore command "Z 0" is added to "clear RIT" command to be sure that also XIT is cleared then.

Wsjt-x controls rig but for reason that was not explained by develop team it does not disable RIT and XIT. RIT is not fatal, but XIT may cause transmits off the used frequency slot and it is hard to notice by user because waterfall TX mark shows still a point inside frequency slot.
That is why command disableRitXit is added when entering to wsjt-remote mode.