ADIF export submode extensions Added all mode/submode enumerations from https://adif.org/312/ADIF_312.htm#Mode_Enumeration to ADIF export tag creation. This way user can add to "modes/user defined digital modes" any submode written in same form as enumeration table and get it exported properly with mode+submode tags.  This addition is random tested as there are so many submodes. How ever I tried to use copy/paste from adif.org for creating source to avoid typing errors in submode names.
Added also modes that come from rig (ic7300) if "auto" mode checkbox is checked and "DATA" is used with any of USB,LSB,FM,AM. They all are now exported with tag PKT.
It may not be the right one if user uses additional program to work with and that program does not have remote link to cqrlog to tell the mode name. In anyway PKT is included in ADIF mode enumerations where modes from rig PKTUSB,PKTLSB,PKTFM and PKTAM are not.

Fixed help (Quick start/modes)