Fix to label printing
If there are qsos that are imported from another program they may not have
QSL manager information. It is not added in adif import process even when cqrlog's
manger database has the information.

Normally this does not show up to user because if the qso is viewed or edited
QSL manager information is fetched from database.
How ever in case of just view qso user may think manager info is ok.
How ever it is saved to qso record only when editing qso.

When user prints QSL labels they are printed without manager info.

This fix checks QSL manager, and if empty, tries to find one from cqrlog's mgr database. If valid mgr is found it is added to label, but not to original qso record.